[refactor] move underscore mixin to shared utilities to keep code DRY
