+++ utilities/shared/shared.js
@@ -2,6 +2,7 @@
 Shared backend utility functions go here
 **************************************************************/
 var fs = require('fs');
+var _ = require('underscore');
 
 /*************************************************************
 Example of writing to a file
@@ -22,3 +23,27 @@ var histogram = function(ar){
     return obj;
   }, {});
 };
+
+/*************************************************************
+Underscore mixin
+**************************************************************/
+
+_.mixin({
+  'sortKeysBy': function (obj, comparator) {
+    var keys = _.sortBy(_.keys(obj), function (key) {
+      return comparator ? comparator(obj[key], key) : key;
+    });
+
+    return _.object(keys, _.map(keys, function (key) {
+      return obj[key];
+    }));
+  }
+});
+
+//returns an object sorted by its keys in descending order
+exports.sortKeysBy = function(obj) {
+    return _.sortKeysBy(obj, function (value, key) {
+            //changes from ascending to descending sort
+            return -(value);
+          })
+}
